add user methods and update migration for location from integer to decimal

+++ db/migrate/20190826171923_add_locations_table.rb
@@ -2,8 +2,8 @@ class AddLocationsTable < ActiveRecord::Migration[6.0]
   def change
     create_table :locations do |t|
       t.column :thing_id, :integer
-      t.column :lat, :integer
-      t.column :long, :integer
+      t.column :lat, :decimal
+      t.column :long, :decimal
     end
   end
 end
